Handle windows filenames.  Initial functioning version.
